feat(Req 06): :sparkles: Implements the function filter_by_job_type() within insights.py
Resolution using the filter() function using a lambda as an argument comparing the job_type key from current iterator position with the job_type provided by the mother function. The main difference I've noted from JS, is that in Python you pass a comparison function as the first argument, the list/array to be filtered as the second argument, while in JS you call filter() as a method and use a callback for making the condition. It's different but also makes sense.